requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
faiss-cpu
streamlit>=1.40.0
python-dotenv>=1.0.0
typing-extensions>=4.9.0
